Story 2614. Moved is_uri_string function to vrm.py.
