+++ vos/vos/commands/vrm.py
@@ -1,14 +1,42 @@
 #!python
 """remove a vospace data node, fails if container node or node is locked."""
 import logging
+import six.moves.urllib.parse as urllibparse
 from ..commonparser import set_logging_level_from_args, exit_on_exception, \
     CommonParser
 from .. import vos, storage_inventory
+
+urlparse = urllibparse.urlparse
 
 DESCRIPTION = """remove a file or a vospace data node; fails if container
 node or node is locked.
 
 eg. vrm vos:/root/node   -- deletes a data node"""
+
+
+def is_uri_string(id_str):
+    """
+    Takes a file identifier string and determines if it specifies a URI.
+    At time of writing, wild card in path is considered invalid.
+
+    :param id_str A identifier string
+    :return True if we can use the identifier string to create a URI instance,
+            False if the identifier string does not start with '<scheme>:'
+    """
+
+    if id_str is None:
+        raise ValueError('Missing identifier: {}'.format(id_str))
+
+    url = urlparse(id_str)
+    if (len(url.scheme) == 0):
+        # missing scheme
+        return False
+    elif (len(url.netloc) == 0) or (len(url.path) == 0) or \
+            ('*' in url.path):
+        raise ValueError('Invalid URL: {}'.format(id_str))
+    else:
+        # a valid url
+        return True
 
 
 def delete_nodes(args):
@@ -44,7 +72,7 @@ def delete_files(args):
                 '{} is not a valid storage file handle'.format(file))
         elif client.isdir(file):
             raise Exception('{} is a directory'.format(file))
-        elif vos.is_uri_string(file):
+        elif is_uri_string(file):
             logging.info('deleting {}'.format(file))
             client.delete(file)
         else:
